requirements.txt
fastapi
uvicorn[standard]
jinja2
streamlit

transformers

langchain
langchain-core
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
tiktoken
chromadb

unstructured
unstructured[docx]
unstructured[xlsx]
python-docx
olefile
pyhwp
PyMuPDF
olefile

Pillow

python-dotenv

beautifulsoup4
selenium
webdriver-manager
requests

APScheduler